refactor: move SphericalCamera to Camera
